Extract virtual-files plugin for clarity
